Added new greedy clustering approach.
These classes are written to not store any peak lists during clustering. This is not needed since they can only add spectra.
The GreedyIncrementalClusteringEngine can only be used with SimilarityCheckerS for which a CDF is available. The similarity threshold used is the clustering accuracy (inversed maximum incorrectly clustered spectra.)
The appenders and ParserUtilities have been adadapted to support these cases without peak lists.